refactor teampage, remove imports, change image format, dellete unused code
