+++ nba-api-dotnet/dbContext.cs
@@ -28,7 +28,7 @@ public class NBAContext : DbContext
 
         if(IsDevelopment){
             // connect to sql server with connection string from app settings
-            options.UseSqlServer(Configuration.GetConnectionString("DanLaptopDB"));
+            options.UseSqlServer(Configuration.GetConnectionString("EveDesktopDB"));
         }else if(IsStaging){
              options.UseSqlServer(Configuration["AzureStagingDatabase"]);
         }else
