+++ nba-front-end/src/App.css
@@ -13,8 +13,7 @@
 
 .App-header {
   overflow: hidden;
-  height: 200px;
-  margin-top:-40px;
+  height: 170px;
 }
 
 .App-link {
