+++ nba-front-end/src/components/Base.tsx
@@ -1,16 +1,13 @@
-import * as React from 'react';
-import { Link, Outlet } from 'react-router-dom';
-import logo from '../logo.svg';
-import { height } from '@mui/system';
+import { Outlet } from 'react-router-dom';
 
 function BasePage() {
   return (
-    <div>
+    <>
       {/* An <Outlet> renders whatever child route is currently active,
           so you can think about this <Outlet> as a placeholder for
         the child routes we defined above. */}        
       <Outlet/>
-    </div>
+    </>
   );
 }
