+++ nba-front-end/src/components/PlayerPage/PlayerPage.tsx
@@ -1,7 +1,5 @@
-import * as React from 'react';
 import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
-import Paper from '@mui/material/Paper';
 import PlayerTableLoader from './PlayerTableLoader';
 
 function PlayerPage() {
