+++ nba-front-end/src/components/TeamPage/TeamPlayerTable.tsx
@@ -1,16 +1,12 @@
-import * as React from 'react';
-import { DataGrid, GridColDef, GridFilterModel, GridValueGetterParams, GridSelectionModel  } from '@mui/x-data-grid';
+import { DataGrid, GridColDef, GridFilterModel, GridValueGetterParams  } from '@mui/x-data-grid';
 import { FormControl, Grid, InputAdornment, InputLabel, OutlinedInput } from '@mui/material';
 import SearchIcon from '@mui/icons-material/Search';
 import { useEffect, useState } from 'react';
-import { axiosRequestConfiguration } from "../../services/axios_config";
-import axios, { AxiosError } from 'axios';
-import Button from '@mui/material/Button';
 import RemovePlayerButton from './RemovePlayer/RemovePlayerButton';
 import RemovePlayerPopUp from './RemovePlayer/RemovePlayerPopUp';
 
 // Setting up the columns of the player table
-const TeamPlayerTable: React.FC<any> = (props) => {
+function TeamPlayerTable(props:any) {
 
 const teamPlayerColumns: GridColDef[] = [
   {
@@ -66,10 +62,10 @@ const teamPlayerColumns: GridColDef[] = [
     }
 
   // initialise the value for the searchbar
-  const [search, setSearch] = React.useState('');
+  const [search, setSearch] = useState('');
 
   // initialise the parameters that the table uses to filter values (when using the searchbar)
-  const [filterModel, setFilterModel] = React.useState<GridFilterModel>({
+  const [filterModel, setFilterModel] = useState<GridFilterModel>({
     items: [
       {
         columnField: 'FullName',
@@ -78,12 +74,6 @@ const teamPlayerColumns: GridColDef[] = [
       },
     ],
   });
-
-  // when you type in the searchbar, update the value of the object
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setSearch(event.target.value);
-    // can't update anything else here because of how the hook works, use useEffect hook instead
-  }
 
   // when [search] is updated, update the table's filter
   useEffect(() => {
@@ -110,7 +100,7 @@ const teamPlayerColumns: GridColDef[] = [
               id="outlined-search"
               label="Search for a player"
               value={search}
-              onChange={handleChange}
+              onChange={(event)=>setSearch(event.target.value)}
               endAdornment={
                 <InputAdornment position="end">
                   <SearchIcon />
