+++ nba-front-end/src/services/axios_setup.ts
@@ -1,4 +1,4 @@
-import axios, { AxiosRequestConfig, AxiosInstance, AxiosPromise } from 'axios';
+import axios, { AxiosRequestConfig, AxiosInstance } from 'axios';
 
 const initialiseAxios = (config: AxiosRequestConfig): AxiosInstance => {
     const axiosInstance = axios.create(config);
